Moved filter change detection back into filter.component. Added ngOnChanges to input-filter to properly updated on setFilters.
