+++ src/ng2-smart-table/components/filter/default-filter.component.ts
@@ -1,6 +1,5 @@
-import { Component, OnChanges, SimpleChanges } from '@angular/core';
+import {Component, Input} from '@angular/core';
 
-import { Subscription } from 'rxjs';
 import {FilterDefault} from "./filter-default";
 
 @Component({
@@ -34,30 +33,6 @@ import {FilterDefault} from "./filter-default";
     </ng-container>
   `,
 })
-export class DefaultFilterComponent extends FilterDefault implements OnChanges {
-
-  protected dataChangedSub: Subscription;
-
-  ngOnChanges(changes: SimpleChanges) {
-    if (changes.source) {
-      if (!changes.source.firstChange) {
-        this.dataChangedSub.unsubscribe();
-      }
-      this.dataChangedSub = this.source.onChanged().subscribe((dataChanges) => {
-        const filterConf = this.source.getFilter();
-        if (filterConf && filterConf.filters && filterConf.filters.length === 0) {
-          this.query = '';
-
-          // add a check for existing filters an set the query if one exists for this column
-          // this covers instances where the filter is set by user code while maintaining existing functionality
-        } else if (filterConf && filterConf.filters && filterConf.filters.length > 0) {
-          filterConf.filters.forEach((k: any, v: any) => {
-            if (k.field == this.column.id) {
-              this.query = k.search;
-            }
-          });
-        }
-      });
-    }
-  }
+export class DefaultFilterComponent extends FilterDefault {
+  @Input() query: string;
 }
