+++ src/ng2-smart-table/components/filter/filter-types/input-filter.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import {Component, OnChanges, OnInit, SimpleChanges} from '@angular/core';
 import { FormControl } from '@angular/forms';
 import { debounceTime, distinctUntilChanged, skip } from 'rxjs/operators';
 
@@ -15,7 +15,7 @@ import { DefaultFilter } from './default-filter';
       placeholder="{{ column.title }}"/>
   `,
 })
-export class InputFilterComponent extends DefaultFilter implements OnInit {
+export class InputFilterComponent extends DefaultFilter implements OnInit, OnChanges {
 
   inputControl = new FormControl();
 
@@ -37,5 +37,11 @@ export class InputFilterComponent extends DefaultFilter implements OnInit {
         this.query = this.inputControl.value;
         this.setFilter();
       });
+  }
+
+  ngOnChanges(changes: SimpleChanges) {
+    if (changes.query) {
+      this.inputControl.setValue(this.query);
+    }
   }
 }
